+++ elastic-beanstalk-deploy/.platform/hooks/postdeploy/00_nginx_set_basic_auth.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -ev
+echo "running postdeploy"
+date >/tmp/platform-hooks-postdeploy.txt
+sudo docker ps | tee -a "/tmp/platform-hooks-postdeploy.txt" >/dev/null