+++ packages/nginx-gateway/launch.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-htpasswd -b -c /etc/nginx/.htpasswd "$OCPU_USER" "$OCPU_PW"
-nginx -g "daemon off;"