+++ packages/radolan-recent-to-s3/test/ftp-server/entrypoint.sh
@@ -1,9 +1,0 @@
-#!/usr/bin/env bash
-set -eo pipefail
-IFS=$'\n\t'
-
-if [ -z "${1}" ]; then
-  /usr/sbin/vsftpd /etc/vsftpd/vsftpd.conf
-else
-  "${@}"
-fi