+++ terraform/s3-bwb-data/outputs.tf
@@ -1,0 +1,3 @@
+output "s3-bucket-name-upload" {
+  value = "${aws_s3_bucket.uploads.bucket_domain_name}"
+}