+++ packages/opencpu-base/CHANGELOG.md
@@ -3,6 +3,97 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.13.0](https://github.com/technologiestiftung/flusshygiene-opencpu-base/compare/v2.12.0...v2.13.0) (2021-05-18)
+
+**Note:** Version bump only for package @fsb/opencpu-base
+
+
+
+
+
+## [2.12.1](https://github.com/technologiestiftung/flusshygiene-opencpu-base/compare/v2.12.0...v2.12.1) (2021-05-17)
+
+**Note:** Version bump only for package @fsb/opencpu-base
+
+
+
+
+
+# [2.12.0](https://github.com/technologiestiftung/flusshygiene-opencpu-base/compare/v2.11.0...v2.12.0) (2021-05-14)
+
+**Note:** Version bump only for package @fsb/opencpu-base
+
+
+
+
+
+# [2.11.0](https://github.com/technologiestiftung/flusshygiene-opencpu-base/compare/v2.11.0-rc.1...v2.11.0) (2021-02-01)
+
+**Note:** Version bump only for package @fsb/opencpu-base
+
+
+
+
+
+# [2.11.0-rc.1](https://github.com/technologiestiftung/flusshygiene-opencpu-base/compare/v2.11.0-rc...v2.11.0-rc.1) (2021-01-27)
+
+
+### Bug Fixes
+
+* **versions:** use fixed versions for all dependencies ([9bc449d](https://github.com/technologiestiftung/flusshygiene-opencpu-base/commit/9bc449d73a42668da43552f0562b8987aed5a8fd))
+
+
+
+
+
+# [2.11.0-rc](https://github.com/technologiestiftung/flusshygiene-opencpu-base/compare/v2.10.6...v2.11.0-rc) (2021-01-20)
+
+**Note:** Version bump only for package @fsb/opencpu-base
+
+
+
+
+
+# [2.11.0-rc](https://github.com/technologiestiftung/flusshygiene-opencpu-base/compare/v2.10.6...v2.11.0-rc) (2021-01-20)
+
+**Note:** Version bump only for package @fsb/opencpu-base
+
+
+
+
+
+# [2.11.0-rc](https://github.com/technologiestiftung/flusshygiene-opencpu-base/compare/v2.10.6...v2.11.0-rc) (2021-01-20)
+
+**Note:** Version bump only for package @fsb/opencpu-base
+
+
+
+
+
+# [2.11.0-rc](https://github.com/technologiestiftung/flusshygiene-opencpu-base/compare/v2.10.6...v2.11.0-rc) (2021-01-20)
+
+**Note:** Version bump only for package @fsb/opencpu-base
+
+
+
+
+
+## [2.10.6](https://github.com/technologiestiftung/flusshygiene-opencpu-base/compare/v2.10.5...v2.10.6) (2020-12-17)
+
+**Note:** Version bump only for package @fsb/opencpu-base
+
+
+
+
+
+
+
+**Note:** Version bump only for package @fsb/opencpu-base
+
+
+
+
+
 ## [2.10.4](https://github.com/technologiestiftung/flusshygiene-opencpu-base/compare/v2.10.0...v2.10.4) (2020-05-28)
 
 **Note:** Version bump only for package @fsb/opencpu-base